Added objects for workspace component

+++ Quadriga/Database/Tables/tbl_project_workspace.sql
@@ -13,7 +13,7 @@ Modified Date : 05/30/2013
 CREATE TABLE IF NOT EXISTS tbl_project_workspace
 (
   projectid       INT           NOT NULL ,
-  workspaceid     INT           NOT NULL ,
+  workspaceid     BIGINT        NOT NULL ,
   updatedby       VARCHAR(10)   NOT NULL,
   updateddate     TIMESTAMP     NOT NULL,
   createdby       VARCHAR(10)   NOT NULL,
